MAINT: move write_file_or_filename util
